PrecRecallAdj
Changed how precision/recall calculation is handled to be able to use Matlab's built-in functions instead of my insane attempt to calculate by hand.  I think it's essentially doing the same thing but hopefully this reduces any potential mistakes on my part.  I left in the option to load and calculate my %Ovlp to display an annotated file over a loaded call file, but still need to address some of my calculation issues...